prevent merge conflicts - organize by folders
