+++ src/components/EditRecipe/EditRecipe.tsx
@@ -21,13 +21,13 @@ import {
 } from "@chakra-ui/react";
 import { AddIcon, CheckIcon, CloseIcon, MinusIcon } from "@chakra-ui/icons";
 import { useParams, useNavigate } from "react-router-dom";
-import { getSingleRecipe, editSingleRecipe } from "../utils/fetchData";
+import { getSingleRecipe, editSingleRecipe } from "../../utils/fetchData";
 import {
   SavedRecipe,
   SavedIngredient,
   RecipeTag,
   EditedRecipe
-} from "../utils/types";
+} from "../../utils/types";
 import IngredientAmountHandle from "./IngredientAmountHandle";
 
 const EditRecipe = () => {
