+++ src/components/EditRecipe/IngredientAmountHandle.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Input, Flex, FormControl, Select } from "@chakra-ui/react";
-import { SavedIngredient } from "../utils/types";
+import { SavedIngredient } from "../../utils/types";
 
 type IngredientAmountHandleProps = {
   ingredient: SavedIngredient;
