+++ src/components/RecipeAI/RecipeAI.tsx
@@ -11,8 +11,8 @@ import {
   UnorderedList
 } from "@chakra-ui/react";
 import { BsHeart, BsHeartFill } from "react-icons/bs";
-import { AIRecipe } from "../utils/types";
-import { saveRecipe } from "../utils/fetchData";
+import { AIRecipe } from "../../utils/types";
+import { saveRecipe } from "../../utils/fetchData";
 import IngredientList from "./IngredientList";
 import InstructionList from "./InstructionList";
