+++ src/components/SavedRecipes/SavedRecipes.tsx
@@ -10,8 +10,8 @@ import {
   GridItem,
   Text
 } from "@chakra-ui/react";
-import { SavedRecipe, RecipeTag } from "../utils/types";
-import { getRecipe } from "../utils/fetchData";
+import { SavedRecipe, RecipeTag } from "../../utils/types";
+import { getRecipe } from "../../utils/fetchData";
 import SavedRecipesList from "./SavedRecipesList";
 import CategoriesList from "./CategoriesList";
 import { useSearchParams } from "react-router-dom";
